requirements.txt
# AI Agent – Python Dependencies
# Install with: pip install -r requirements.txt

# LLM SDKs (core)
anthropic>=0.25.0
openai>=1.30.0

# CLI framework
click>=8.1.0

# Dashboard server
flask>=3.1.3
flask-cors>=6.0.2

# Config parsing
pyyaml>=6.0

# Rich terminal output
rich>=13.0.0

# Testing
pytest>=9.0.2
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0   # load .env files automatically
requests>=2.31.0       # HTTP client

# ── Optional: Cloud Provider SDKs ─────────────────────────────────────────
# Uncomment the ones you need:
#
# AWS Bedrock (Claude in your AWS account):
#   boto3>=1.34.0
#
# GCP Vertex AI (Claude + Gemini in your GCP):
#   anthropic[vertex]>=0.25.0
#   google-cloud-aiplatform>=1.40.0
